requirements.txt
playwright==1.41.0
beautifulsoup4==4.12.3
requests==2.31.0
Pillow==10.2.0
lxml==5.1.0
